requirements.txt
streamlit
openai
python-dotenv
pandas
pydantic
PyMuPDF
python-docx
reportlab